update: Slackログインのロジックをuserモデルから切り出し
